+++ crd/error.go
@@ -10,17 +10,3 @@ var invalidConfigError = microerror.New("invalid config")
 func IsInvalidConfig(err error) bool {
 	return microerror.Cause(err) == invalidConfigError
 }
-
-var nameConflictError = microerror.New("name conflict")
-
-// IsNameConflict asserts nameConflictError.
-func IsNameConflict(err error) bool {
-	return microerror.Cause(err) == nameConflictError
-}
-
-var notEstablishedError = microerror.New("not established")
-
-// IsNotEstablished asserts notEstablishedError.
-func IsNotEstablished(err error) bool {
-	return microerror.Cause(err) == notEstablishedError
-}
